refactor: `BankItem` 컴포저블 패키지 이동
