+++ app/src/main/java/woowacourse/payments/ui/newcard/NewCardScreen.kt
@@ -11,6 +11,7 @@ import androidx.compose.material3.rememberModalBottomSheetState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
@@ -37,7 +38,7 @@ import woowacourse.payments.ui.newcard.dialog.BankBottomSheet
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun NewCardScreen(
-    newCardStateHolder: NewCardStateHolder = rememberSaveable { NewCardStateHolder() },
+    newCardStateHolder: NewCardStateHolder = remember { NewCardStateHolder() },
     onBackPress: () -> Unit = {},
     onSaved: (PaymentCardUiModel) -> Unit = {},
 ) {
